+++ laxar.js
@@ -3,24 +3,17 @@
  * Released under the MIT license.
  * http://laxarjs.org/license
  */
-import log from './lib/logging/log';
-import * as eventBus from './lib/event_bus/event_bus';
-import * as fileResourceProvider from './lib/file_resource_provider/file_resource_provider';
-import * as i18n from './lib/i18n/i18n';
-import * as widgetLoader from './lib/loaders/widget_loader';
 import assert from './lib/utilities/assert';
-import * as configuration from './lib/utilities/configuration';
 import fn from './lib/utilities/fn';
 import * as object from './lib/utilities/object';
-import * as path from './lib/utilities/path';
-import storage from './lib/utilities/storage';
 import * as string from './lib/utilities/string';
-import { create as createServices } from './lib/runtime/services';
-import * as controlsService from './lib/runtime/controls_service';
-import * as themeManager from './lib/runtime/theme_manager';
-import * as plainAdapter from './lib/widget_adapters/plain_adapter';
-import pageToolingApi from './lib/tooling/pages';
 
+import { create as createServices } from './lib/runtime/services';
+import { createLogErrorHandler } from './lib/event_bus/event_bus';
+import { create as createConfiguration } from './lib/runtime/configuration';
+import { create as createBrowser } from './lib/runtime/browser';
+import { create as createLog } from './lib/logging/log';
+import * as plainAdapter from './lib/widget_adapters/plain_adapter';
 
 /**
  * Bootstraps AngularJS on the provided `anchorElement` and sets up the LaxarJS runtime.
@@ -28,17 +21,37 @@ import pageToolingApi from './lib/tooling/pages';
  * @memberOf laxar
  *
  * @param {HTMLElement} anchorElement the element to insert the page in
+ * @param {Object} optionalOptions
  */
-export function bootstrap( anchorElement, { widgetAdapters, widgetModules } ) {
+export function bootstrap( anchorElement, { widgetAdapters, widgetModules, whenServicesReady, configuration } ) {
 
-   setInstanceIdLogTag();
+   const services = createServices( configuration || {} );
 
-   log.trace( 'Bootstrapping LaxarJS...' );
+   const { globalEventBus, log, i18n, themeManager, cssLoader, paths, storage, widgetLoader } = services;
+   themeManager.loadThemeCss( cssLoader, paths );
+   const adapterModules = [ plainAdapter, ...widgetAdapters ];
+   const adapters = bootstrapWidgetAdapters( services, adapterModules, widgetModules );
+   widgetLoader.registerWidgetAdapters( adapters );
 
-   const adapters = [ plainAdapter, ...widgetAdapters ];
-   const services = createServices( configuration, adapters );
-   loadThemeCss( services );
-   bootstrapWidgetAdapters( services, adapters, widgetModules );
+   // TODO (#310) define exact set of publicy visible services, plus naming (injection-style or module-style?)
+   const publicServices = {
+      configuration: services.configuration,
+      globalEventBus,
+      i18n,
+      log,
+      storage,
+      tooling: services.toolingProviders
+   };
+
+   fallbackLog = log;
+
+   // TODO (#310) move out into application space (`whenServicesReady` callback)?
+   globalEventBus.setErrorHandler( createLogErrorHandler( log ) );
+   // TODO (#310) move out into application space (`whenServicesReady` callback)?
+   ensureInstanceId( log, storage );
+   if( whenServicesReady ) {
+      whenServicesReady( publicServices );
+   }
 
    whenDocumentReady( () => {
       log.trace( `Loading flow from "${services.paths.FLOW_JSON}"` );
@@ -54,48 +67,6 @@ export function bootstrap( anchorElement, { widgetAdapters, widgetModules } ) {
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-function setInstanceIdLogTag() {
-   const instanceIdStorageKey = 'axLogTags.INST';
-   const store = storage.getApplicationSessionStorage();
-   let instanceId = store.getItem( instanceIdStorageKey );
-   if( !instanceId ) {
-      instanceId = '' + Date.now() + Math.floor( Math.random() * 100 );
-      store.setItem( instanceIdStorageKey, instanceId );
-   }
-
-   log.addTag( 'INST', instanceId );
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-function loadThemeCss( services ) {
-   services.themeManager
-      .urlProvider( path.join( services.paths.THEMES, '[theme]' ), null, [ services.paths.DEFAULT_THEME ] )
-      .provide( [ 'css/theme.css' ] )
-      .then( ( [ cssFile ] ) => services.cssLoader.load( cssFile ) );
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
-function bootstrapWidgetAdapters( services, adapters, widgetModules ) {
-   const widgetAdapters = {};
-   adapters.forEach( adapter => {
-      widgetAdapters[ adapter.technology ] = adapter;
-   } );
-
-   Object.keys( widgetModules ).forEach( technology => {
-      const adapter = widgetAdapters[ technology ];
-      if( !adapter ) {
-         log.fatal( 'Unknown widget technology: [0]', technology );
-         return;
-      }
-
-      adapter.bootstrap( widgetModules[ technology ], services );
-   } );
-}
-
-//////////////////////////////////////////////////////////////////////////////////////////////////////////////
-
 function whenDocumentReady( callback ) {
    if( document.readyState === 'complete' ) {
       callback();
@@ -105,31 +76,72 @@ function whenDocumentReady( callback ) {
    }
 }
 
+///////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+function bootstrapWidgetAdapters( services, adapterModules, widgetModulesByTechnology ) {
+   const { log } = services;
+   const adapterModulesByTechnology = {};
+   adapterModules.forEach( module => {
+      adapterModulesByTechnology[ module.technology ] = module;
+   } );
+
+   const adapters = [];
+   Object.keys( widgetModulesByTechnology ).forEach( technology => {
+      const adapterModule = adapterModulesByTechnology[ technology ];
+      if( !adapterModule ) {
+         log.fatal( 'Unknown widget technology: [0]', technology );
+         return;
+      }
+      const widgetModules = widgetModulesByTechnology[ technology ];
+      adapters.push( adapterModule.bootstrap( widgetModules, services ) );
+   } );
+   return adapters;
+}
+
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-// API to leverage tooling support. Not for direct use by widgets/activities!
-// For example laxar-mocks needs this for widget tests
-const _tooling = {
-   controlsService: controlsService,
-   eventBus: eventBus,
-   fileResourceProvider: fileResourceProvider,
-   path: path,
-   themeManager: themeManager,
-   widgetLoader: widgetLoader,
-   // Prototype support for page inspection tools:
-   pages: pageToolingApi
-};
+function ensureInstanceId( log, storage ) {
+   const instanceIdStorageKey = 'axLogTags.INST';
+   const store = storage.getApplicationSessionStorage();
+   let instanceId = store.getItem( instanceIdStorageKey );
+   if( !instanceId ) {
+      instanceId = '' + Date.now() + Math.floor( Math.random() * 100 );
+      store.setItem( instanceIdStorageKey, instanceId );
+   }
+   log.addTag( 'INST', instanceId );
+}
+
+//////////////////////////////////////////////////////////////////////////////////////////////////////////////
+
+// Stores the fallback logger. The initial log is replaced with a correctly configured instance as soon as
+// the laxarjs services have been bootstrapped.
+let fallbackLog = createLog(
+   createConfiguration( { logging: { level: 'INFO' } } ),
+   createBrowser()
+);
+
+/**
+ * To ease the transition from laxarjs v1 to laxarjs v2, a global log fallback is provided.
+ * Clients should prefer the widget-level injection `axLog` (TODO, #306) or use the global log-service,
+ * which can be obtained using the `whenServicesReady` callback to `bootstrap`.
+ */
+const log = object.tabulate(
+   method => {
+      return function() {
+         // TODO, #306: enable this deprecation warning
+         // fallbackLog.warn( 'Deprecation warning: avoid using laxar.log and prefer the axLog injection' );
+         fallbackLog[ method ].apply( fallbackLog, arguments );
+      };
+   },
+   Object.keys( fallbackLog )
+);
 
 //////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
 export {
    assert,
-   configuration,
-   fn,
-   i18n,
-   log,
    object,
-   storage,
    string,
-   _tooling
+   fn,
+   log
 };