+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Last Changes
 
+
+## v2.0.0-alpha.1
+
 - [#303](https://github.com/LaxarJS/laxar/issues/298): modules: eliminated all global state
     + **BREAKING CHANGE:** see ticket for details
 - [#298](https://github.com/LaxarJS/laxar/issues/298): modules: load using webpack